requirements.txt
matplotlib==3.5.3
fastai==2.7.10
numpy==1.21.6
opencv_python==4.6.0.66
openslide_python==1.2.0
pandas==1.1.5
seaborn==0.12.2
Shapely==1.8.5.post1
pytorch-cuda==11.8
tqdm==4.64.1
timm==1.0.15
transformers==4.51.2
